safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.1_The_do_while_Statement/S12.6.1_A6_T6-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected string literal 'hood'. Expected '(' to start a do-while loop condition.
S12.6.1_A6_T6

PASS Expected parsing failure

TEST COMPLETE
